current-user

+++ app/controllers/shifts_controller.rb
@@ -1,26 +1,37 @@
 class ShiftsController < ApplicationController
   skip_before_action :require_login, only: [:create_schedule]
   skip_before_action :verify_authenticity_token, only: [:create_schedule]
-  
+
   def index
     Rails.logger.debug "Calling set_month_data in ShiftsController#index"
     set_month_data(params[:year], params[:month])
-    @employees = Employee.all
-    @shifts = Shift.all.group_by(&:employee_id)
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
+    
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
+
+    # current_userに関連するシフトデータを取得
+    @shifts = Shift.where(employee_id: @employees.pluck(:id)).group_by(&:employee_id)
+
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
+    
     @shift_counts = sum_shift_counts(@employees, @shifts, @calendar)
   end
 
   def new
     @year = params[:year]
     @month = params[:month]
-    
-    set_month_data(@year, @month)
-    @employees = Employee.all
 
-    # ここでシフトリクエストを取得します
-    @shift_requests = ShiftRequest.where(date: @start_date..@end_date).group_by(&:employee_id)
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
+    set_month_data(@year, @month)
+
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
+
+    # current_userに関連するシフトリクエストを取得
+    @shift_requests = ShiftRequest.where(date: @start_date..@end_date, user_id: current_user.id).group_by(&:employee_id)
+    
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
   end
 
   def edit
@@ -30,10 +41,17 @@ class ShiftsController < ApplicationController
 
   def update_schedule
     set_month_data(params[:year], params[:month])
-    @employees = Employee.all
+
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
+
     @schedule_output = @schedule_output || {}
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
-    @shifts = Shift.where(date: @start_date..@end_date).group_by(&:employee_id)
+
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
+
+    # current_userに関連するシフトデータを取得
+    @shifts = Shift.where(date: @start_date..@end_date, employee_id: @employees.pluck(:id)).group_by(&:employee_id)
     
     ActiveRecord::Base.transaction do
       params[:shifts].each do |employee_id, shifts|
@@ -49,7 +67,7 @@ class ShiftsController < ApplicationController
       end
     
       params[:memos].each do |date, content|
-        memo = Memo.find_by(date: date)
+        memo = Memo.find_by(date: date, user_id: current_user.id)
         if content.blank?
           memo&.destroy
         else
@@ -82,12 +100,16 @@ class ShiftsController < ApplicationController
   def create_schedule
     Rails.logger.debug "Calling set_month_data in ShiftsController#create_schedule"
     set_month_data(params[:year], params[:month]) 
-    @employees = Employee.all
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
+    
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
 
-    employees = Employee.all
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
+
+    employees = current_user.employees
     dates = (@start_date..@end_date).to_a
-    shift_requests = ShiftRequest.where(date: @start_date..@end_date).group_by(&:employee_id)
+    shift_requests = ShiftRequest.where(date: @start_date..@end_date, user_id: current_user.id).group_by(&:employee_id)
 
     data = {
       employees: employees.map do |e|
@@ -139,27 +161,33 @@ class ShiftsController < ApplicationController
 
   def edit_schedule
     set_month_data(params[:year], params[:month])
-    @employees = Employee.all
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
-    @shifts = Shift.where(date: @start_date..@end_date).group_by(&:employee_id)
+
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
+
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
+
+    # current_userに関連するシフトデータを取得
+    @shifts = Shift.where(date: @start_date..@end_date, employee_id: @employees.pluck(:id)).group_by(&:employee_id)
     @shifts.default = [] 
   end
 
   def destroy_all
     start_date = Date.new(params[:year].to_i, params[:month].to_i, 1)
     end_date = start_date.end_of_month
-  
+
     ActiveRecord::Base.transaction do
-      # シフトを削除
-      Shift.where(date: start_date..end_date).destroy_all
+      # current_userに関連するシフトを削除
+      Shift.where(date: start_date..end_date, employee_id: current_user.employees.pluck(:id)).destroy_all
       
-      # シフトリクエストを削除
-      ShiftRequest.where(date: start_date..end_date).destroy_all
+      # current_userに関連するシフトリクエストを削除
+      ShiftRequest.where(date: start_date..end_date, user_id: current_user.id).destroy_all
       
-      # メモを削除
-      Memo.where(date: start_date..end_date).destroy_all
+      # current_userに関連するメモを削除
+      Memo.where(date: start_date..end_date, user_id: current_user.id).destroy_all
     end
-    
+
     redirect_to shifts_path(year: params[:year], month: params[:month]), success: 'シフトを削除しました'
   rescue ActiveRecord::RecordNotDestroyed => e
     redirect_to shifts_path(year: params[:year], month: params[:month]), alert: "削除に失敗しました: #{e.message}"
@@ -167,9 +195,15 @@ class ShiftsController < ApplicationController
 
   def export_excel
     set_month_data(params[:year], params[:month])
-    @employees = Employee.all
-    @shifts = Shift.where(date: @start_date..@end_date).group_by(&:employee_id)
-    @memos = Memo.where(date: @start_date..@end_date).index_by(&:date)
+
+    # current_userに関連する従業員を取得
+    @employees = current_user.employees
+
+    # current_userに関連するシフトデータを取得
+    @shifts = Shift.where(date: @start_date..@end_date, employee_id: @employees.pluck(:id)).group_by(&:employee_id)
+
+    # current_userに関連するメモを取得
+    @memos = Memo.where(date: @start_date..@end_date, user_id: current_user.id).index_by(&:date)
   
     respond_to do |format|
       format.xlsx {
@@ -222,7 +256,7 @@ class ShiftsController < ApplicationController
   end
 
   def clear_shift_requests_and_memos(start_date, end_date)
-    ShiftRequest.where(date: start_date..end_date).destroy_all
-    Memo.where(date: start_date..end_date).destroy_all
+    ShiftRequest.where(date: start_date..end_date, user_id: current_user.id).destroy_all
+    Memo.where(date: start_date..end_date, user_id: current_user.id).destroy_all
   end
 end
